Extract ServiceClient class
